use int64_t for addresses

+++ nanocube.h
@@ -42,14 +42,14 @@ struct Nanocube {
   typedef std::unordered_map<pair<int, int>, int, PairHasher> SummaryCache;
 
   pair<int, int> insert_fresh_node
-  (const Summary &summary, const vector<int> &addresses, int dim, int bit);
+  (const Summary &summary, const vector<int64_t> &addresses, int dim, int bit);
   
   pair<int, int> insert_node
-  (const Summary &summary, const vector<int> &addresses, int current_node, int current_dim, int current_bit, SummaryCache &summary_cache);
+  (const Summary &summary, const vector<int64_t> &addresses, int current_node, int current_dim, int current_bit, SummaryCache &summary_cache);
 
   pair<int, int> merge(int left, int right, int dim, SummaryCache &summary_cache);
 
-  void insert(const Summary &summary, const vector<int> &addresses);
+  void insert(const Summary &summary, const vector<int64_t> &addresses);
 
   /****************************************************************************/
   // simple accessors
